Harden gateway security and split API modules
Tightens security across API and mobile flows by failing closed when client IP is unknown, validating Cybara deep-link routes, and enforcing an external URL allowlist before opening links. Refactors large surfaces for maintainability by extracting the macOS Tasks screen, splitting web API clients into focused modules (wallet/lab/computer-use/shared client), and moving TUI chat/provider logic into dedicated helpers. CLI gateway/process handling now uses Bun-native process management with structured API errors and better long-lived key handling, with broad test updates covering regressions and new behavior.
