+++ apps/mobile/src/lib/externalUrlPolicy.ts
@@ -1,0 +1,9 @@
+const ALLOWED_EXTERNAL_PROTOCOLS = new Set(["http:", "https:", "mailto:", "tel:"]);
+
+export function isAllowedExternalUrl(value: string): boolean {
+  try {
+    return ALLOWED_EXTERNAL_PROTOCOLS.has(new URL(value).protocol.toLowerCase());
+  } catch {
+    return false;
+  }
+}
