+++ eslint.config.js
@@ -48,7 +48,8 @@ export default [
         rules: {
             "no-unused-vars": "off",
             "no-undef": "off",
-            "@typescript-eslint/no-explicit-any": "warn",
+            "@typescript-eslint/no-explicit-any": "error",
+            "@typescript-eslint/no-non-null-assertion": "error",
             "@typescript-eslint/no-unused-vars": ["warn", { argsIgnorePattern: "^_" }],
             "prefer-const": "error",
             "no-var": "error",
