+++ src/cli/client.ts
@@ -36,6 +36,25 @@ export function resolveCliGatewayPassword(
 
 export const CLI_GATEWAY_PASSWORD = resolveCliGatewayPassword();
 
+export class CliApiError extends Error {
+  readonly endpoint: string;
+  readonly status: number | null;
+
+  constructor(endpoint: string, message: string, status: number | null = null) {
+    super(message);
+    this.name = "CliApiError";
+    this.endpoint = endpoint;
+    this.status = status;
+  }
+}
+
+export interface CliApiRequestContext {
+  apiBase?: string;
+  apiKey?: string | null;
+  gatewayPassword?: string | null;
+  fetchImpl?: typeof fetch;
+}
+
 export function buildCliAuthHeaders(
   apiKey: string | null,
   headers?: RequestInit["headers"],
@@ -59,29 +78,88 @@ export function withCliAuthHeaders(
   headers?: RequestInit["headers"],
   ensureJsonContentType = false
 ): Headers {
-  return buildCliAuthHeaders(CLI_API_KEY, headers, ensureJsonContentType, CLI_GATEWAY_PASSWORD);
+  return buildCliAuthHeaders(
+    resolveCliApiKey(),
+    headers,
+    ensureJsonContentType,
+    resolveCliGatewayPassword()
+  );
+}
+
+function apiErrorDetail(value: unknown): string {
+  if (!value || typeof value !== "object" || Array.isArray(value)) return "";
+  const record = value as Record<string, unknown>;
+  for (const key of ["message", "error", "detail"]) {
+    const detail = record[key];
+    if (typeof detail === "string" && detail.trim()) return detail.trim();
+  }
+  return "";
+}
+
+async function responseErrorDetail(response: Response): Promise<string> {
+  const text = await response.text().catch(() => "");
+  if (!text.trim()) return "";
+  try {
+    return apiErrorDetail(JSON.parse(text)) || text.trim();
+  } catch {
+    return text.trim();
+  }
+}
+
+export async function requestCliAPI<T>(
+  endpoint: string,
+  options?: RequestInit,
+  context: CliApiRequestContext = {}
+): Promise<T> {
+  const apiBase = context.apiBase ?? CLI_API_BASE;
+  const apiKey = context.apiKey === undefined ? resolveCliApiKey() : context.apiKey;
+  const gatewayPassword =
+    context.gatewayPassword === undefined ? resolveCliGatewayPassword() : context.gatewayPassword;
+  const fetchImpl = context.fetchImpl ?? fetch;
+  let response: Response;
+  try {
+    response = await fetchImpl(`${apiBase}${endpoint}`, {
+      ...options,
+      headers: buildCliAuthHeaders(apiKey, options?.headers, true, gatewayPassword),
+    });
+  } catch (error) {
+    const message = error instanceof Error ? error.message : String(error);
+    throw new CliApiError(endpoint, message);
+  }
+  if (!response.ok) {
+    const detail = await responseErrorDetail(response);
+    const statusText = response.statusText.trim();
+    const suffix = detail || statusText || "Request failed";
+    throw new CliApiError(endpoint, `HTTP ${response.status}: ${suffix}`, response.status);
+  }
+  try {
+    return (await response.json()) as T;
+  } catch {
+    throw new CliApiError(endpoint, "Gateway returned invalid JSON", response.status);
+  }
+}
+
+export function formatCliApiError(error: unknown, apiBase = CLI_API_BASE): string {
+  const message = error instanceof Error ? error.message : String(error);
+  const endpoint = error instanceof CliApiError ? error.endpoint : "";
+  const target = endpoint ? `${apiBase}${endpoint}` : apiBase;
+  if (error instanceof CliApiError && error.status === 401) {
+    return `Unauthorized (401) from ${target}. Refresh the local API key or CYBARA_API_KEY.`;
+  }
+  if (error instanceof CliApiError && error.status === 403) {
+    return `Access denied (403) from ${target}. Set CYBARA_GATEWAY_PASSWORD if required.`;
+  }
+  if (message.includes("ECONNREFUSED") || message.includes("fetch failed")) {
+    return `Cannot connect to Cybara at ${apiBase}. Start it with: cybara start`;
+  }
+  return endpoint ? `${endpoint}: ${message}` : message;
 }
 
 export async function fetchCliAPI<T>(endpoint: string, options?: RequestInit): Promise<T | null> {
   try {
-    const response = await fetch(`${CLI_API_BASE}${endpoint}`, {
-      ...options,
-      headers: withCliAuthHeaders(options?.headers, true),
-    });
-    if (!response.ok) throw new Error(`HTTP ${response.status} ${response.statusText}`);
-    return (await response.json()) as T;
+    return await requestCliAPI<T>(endpoint, options);
   } catch (error) {
-    const message = error instanceof Error ? error.message : String(error);
-    if (message.includes("ECONNREFUSED") || message.includes("fetch failed")) {
-      console.error(`ERROR: Cannot connect to Cybara at ${CLI_API_BASE}`);
-      console.error("Is the server running? Start it with: cybara start");
-    } else if (message.includes("HTTP 401")) {
-      console.error("ERROR: Unauthorized API request (401)");
-      console.error("Set CYBARA_API_KEY or create ~/.cybara/api_key");
-    } else if (message.includes("HTTP 403")) {
-      console.error("ERROR: Gateway access denied (403)");
-      console.error("Set CYBARA_GATEWAY_PASSWORD when the remote gateway requires one");
-    }
+    console.error(`ERROR: ${formatCliApiError(error)}`);
     return null;
   }
 }
