+++ tests/cli/client-contracts.test.ts
@@ -4,8 +4,12 @@ import { tmpdir } from "os";
 import { join } from "path";
 import {
   buildCliAuthHeaders,
+  CliApiError,
+  formatCliApiError,
+  requestCliAPI,
   resolveCliApiKey,
   resolveCliGatewayPassword,
+  withCliAuthHeaders,
 } from "../../src/cli/client";
 import {
   type AgentItem,
@@ -15,6 +19,8 @@ import {
 } from "../../src/cli/contracts";
 
 const tempDirectories: string[] = [];
+const originalCybaraHome = process.env.CYBARA_HOME;
+const originalCybaraApiKey = process.env.CYBARA_API_KEY;
 
 function createTempHome(): string {
   const directory = mkdtempSync(join(tmpdir(), "cybara-cli-client-"));
@@ -23,6 +29,10 @@ function createTempHome(): string {
 }
 
 afterEach(() => {
+  if (originalCybaraHome === undefined) delete process.env.CYBARA_HOME;
+  else process.env.CYBARA_HOME = originalCybaraHome;
+  if (originalCybaraApiKey === undefined) delete process.env.CYBARA_API_KEY;
+  else process.env.CYBARA_API_KEY = originalCybaraApiKey;
   for (const directory of tempDirectories.splice(0)) {
     rmSync(directory, { recursive: true, force: true });
   }
@@ -77,6 +87,42 @@ describe("CLI client contracts", () => {
       "gateway-secret"
     );
     expect(preservedPassword.get("X-Cybara-Gateway-Password")).toBe("override");
+  });
+
+  test("reloads the local API key for long-lived CLI and TUI requests", () => {
+    const home = createTempHome();
+    process.env.CYBARA_HOME = home;
+    delete process.env.CYBARA_API_KEY;
+    writeFileSync(join(home, "api_key"), "first-key\n");
+    expect(withCliAuthHeaders().get("Authorization")).toBe("Bearer first-key");
+    writeFileSync(join(home, "api_key"), "rotated-key\n");
+    expect(withCliAuthHeaders().get("Authorization")).toBe("Bearer rotated-key");
+  });
+
+  test("preserves endpoint and status details for TUI error rendering", async () => {
+    const request = requestCliAPI("/api/providers", undefined, {
+      apiBase: "http://127.0.0.1:4269",
+      apiKey: "stale-key",
+      fetchImpl: async () =>
+        new Response(JSON.stringify({ error: "Invalid API key" }), {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        }),
+    });
+    const error = await request.catch((cause: unknown) => cause);
+    expect(error).toBeInstanceOf(CliApiError);
+    expect((error as CliApiError).endpoint).toBe("/api/providers");
+    expect((error as CliApiError).status).toBe(401);
+    expect(formatCliApiError(error)).toContain("Unauthorized (401)");
+    expect(formatCliApiError(error)).toContain("/api/providers");
+  });
+
+  test("rejects malformed successful gateway responses", async () => {
+    await expect(
+      requestCliAPI("/api/providers", undefined, {
+        fetchImpl: async () => new Response("not-json", { status: 200 }),
+      })
+    ).rejects.toThrow("Gateway returned invalid JSON");
   });
 });
