+++ ui/src/lib/api.ts
@@ -1,4 +1,4 @@
-import { apiFetch } from "@/lib/auth";
+import { fetchApi } from "@/lib/api-client";
 import type { PendingChatMessage } from "@/lib/status-stream";
 import type {
   Agent,
@@ -23,8 +23,9 @@ import type {
   Skill,
   Task,
 } from "@/types";
+import type { AgentGolden } from "@/lib/api/lab";
 
-const API_BASE = "/api";
+export { extractApiError } from "@/lib/api-client";
 
 type ChatProcessActivityPayload = Array<{
   id?: string;
@@ -113,36 +114,6 @@ export interface NearbyStatus {
     messageCount: number;
     workspace: { name: string; branch?: string; commit?: string; dirty?: boolean } | null;
   }>;
-}
-
-/**
- * Resolve a human-readable error from an ApiResponse: prefer a nested
- * `data.error` (envelope returned with HTTP 200), then the transport `error`,
- * then the caller's fallback. Centralizes the chain repeated across hooks.
- */
-export function extractApiError<T>(response: ApiResponse<T>, fallback: string): string {
-  const data = response.data as { error?: unknown } | undefined;
-  const dataError =
-    data && typeof data === "object" && typeof data.error === "string" ? data.error : null;
-  return dataError || response.error || fallback;
-}
-
-async function fetchApi<T>(endpoint: string, options?: RequestInit): Promise<ApiResponse<T>> {
-  const url = `${API_BASE}${endpoint}`;
-  const response = await apiFetch(url, {
-    headers: {
-      "Content-Type": "application/json",
-    },
-    ...options,
-  });
-
-  if (!response.ok) {
-    const error = await response.text();
-    return { success: false, error };
-  }
-
-  const data = await response.json();
-  return { success: true, data };
 }
 
 function asRecord(value: unknown): Record<string, unknown> {
@@ -988,529 +959,7 @@ export const setupApi = {
   complete: () => fetchApi<{ success: boolean }>("/setup/complete", { method: "POST" }),
 };
 
-export interface WalletStatus {
-  exists: boolean;
-  unlocked: boolean;
-  address?: string;
-  createdAt?: string;
-  updatedAt?: string;
-  unlockExpiresAt?: string;
-  wordCount?: number;
-  kdf?: {
-    name: "PBKDF2";
-    hash: "SHA-256";
-    iterations: number;
-  };
-  agentAccessEnabled: boolean;
-  chains: Array<WalletChain>;
-  primaryAddresses?: Record<WalletChain, string>;
-}
-
-export type WalletChain = "eth" | "btc" | "sol";
-export type WalletTokenChain = "eth" | "sol";
-
-export interface WalletAccount {
-  chain: WalletChain;
-  index: number;
-  path: string;
-  address: string;
-}
-
-export interface WalletBalance extends WalletAccount {
-  symbol: "ETH" | "BTC" | "SOL";
-  decimals: number;
-  amount: string;
-  raw: string;
-}
-
-export interface WalletTransaction {
-  chain: WalletChain;
-  txid: string;
-  status: "confirmed" | "pending" | "failed";
-  from?: string;
-  to?: string;
-  amount?: string;
-  fee?: string;
-  confirmations?: number;
-  timestamp?: string;
-  explorerUrl: string;
-}
-
-export interface WalletTokenBalance {
-  chain: WalletTokenChain;
-  index: number;
-  address: string;
-  tokenAddress: string;
-  symbol: string;
-  name?: string;
-  decimals: number;
-  amount: string;
-  raw: string;
-  tokenAccount?: string;
-}
-
-export interface WalletInstructionAccount {
-  pubkey: string;
-  isSigner?: boolean;
-  isWritable?: boolean;
-}
-
-export interface WalletRpcConfig {
-  ethRpc: string;
-  solRpc: string;
-  btcApi: string;
-}
-
-export interface WalletRpcServiceStatus {
-  chain: WalletChain;
-  endpoint: string;
-  healthy: boolean;
-  latencyMs: number;
-  latestHeight?: string;
-  error?: string;
-}
-
-export interface WalletRpcStatus {
-  checkedAt: string;
-  services: WalletRpcServiceStatus[];
-}
-
-export interface WalletTokenTransaction {
-  chain: WalletTokenChain;
-  index: number;
-  address: string;
-  tokenAddress: string;
-  symbol: string;
-  name?: string;
-  decimals: number;
-  txid: string;
-  status: "confirmed" | "pending" | "failed";
-  direction: "in" | "out" | "self" | "unknown";
-  from?: string;
-  to?: string;
-  amount: string;
-  raw: string;
-  fee?: string;
-  timestamp?: string;
-  explorerUrl: string;
-}
-
-export interface WalletAgentPolicy {
-  allowNativeSend: boolean;
-  allowTokenSend: boolean;
-  allowEthContractWrite: boolean;
-  allowSolProgramInstruction: boolean;
-  allowEthSwaps: boolean;
-  allowDappInteraction: boolean;
-  allowX402Payments: boolean;
-  allowedEthContracts: string[];
-  allowedSolPrograms: string[];
-  allowedDappHosts: string[];
-  allowedX402Networks: string[];
-  x402MaxAmountAtomic: string;
-  allowedSendRecipients: string[];
-  maxSendAmount: string;
-}
-
-export interface WalletSwapEthUniswapResult {
-  chain: "eth";
-  dex: "uniswap_v2";
-  from: string;
-  toTokenAddress: string;
-  toTokenSymbol: string;
-  amountInEth: string;
-  amountInWei: string;
-  quotedAmountOut: string;
-  quotedAmountOutRaw: string;
-  minAmountOut: string;
-  minAmountOutRaw: string;
-  slippageBps: number;
-  recipient: string;
-  deadline: string;
-  txid?: string;
-  explorerUrl?: string;
-  dryRun: boolean;
-}
-
-export const walletApi = {
-  status: () => fetchApi<WalletStatus>("/wallet/status"),
-  rpc: () => fetchApi<WalletRpcConfig>("/wallet/rpc"),
-  rpcStatus: () => fetchApi<WalletRpcStatus>("/wallet/rpc/status"),
-  updateRpc: (payload: { ethRpc?: string; solRpc?: string; btcApi?: string }) =>
-    fetchApi<{ success: boolean; config: WalletRpcConfig }>("/wallet/rpc", {
-      method: "PUT",
-      body: JSON.stringify(payload),
-    }),
-  create: (password: string) =>
-    fetchApi<{
-      success: boolean;
-      mnemonic: string;
-      address: string;
-      primaryAddresses: Record<WalletChain, string>;
-    }>("/wallet/create", {
-      method: "POST",
-      body: JSON.stringify({ password }),
-    }),
-  importWallet: (mnemonic: string, password: string) =>
-    fetchApi<{
-      success: boolean;
-      mnemonic: string;
-      address: string;
-      primaryAddresses: Record<WalletChain, string>;
-    }>("/wallet/import", {
-      method: "POST",
-      body: JSON.stringify({ mnemonic, password }),
-    }),
-  unlock: (password: string) =>
-    fetchApi<{
-      success: boolean;
-      address: string;
-      primaryAddresses: Record<WalletChain, string>;
-      unlockExpiresAt: string;
-    }>("/wallet/unlock", {
-      method: "POST",
-      body: JSON.stringify({ password }),
-    }),
-  revealSeed: (password: string, acknowledgement: string) =>
-    fetchApi<{ mnemonic: string; wordCount: number }>("/wallet/seed", {
-      method: "POST",
-      body: JSON.stringify({ password, acknowledgement }),
-    }),
-  lock: () => fetchApi<{ success: boolean }>("/wallet/lock", { method: "POST" }),
-  accounts: (params?: { chains?: WalletChain[]; count?: number; startIndex?: number }) => {
-    const query = new URLSearchParams();
-    if (params?.chains?.length) query.set("chains", params.chains.join(","));
-    if (typeof params?.count === "number") query.set("count", String(params.count));
-    if (typeof params?.startIndex === "number") query.set("startIndex", String(params.startIndex));
-    return fetchApi<WalletAccount[]>(`/wallet/accounts${query.size ? `?${query.toString()}` : ""}`);
-  },
-  receive: (chain: WalletChain, index = 0) =>
-    fetchApi<WalletAccount>(
-      `/wallet/receive?chain=${encodeURIComponent(chain)}&index=${encodeURIComponent(index)}`
-    ),
-  balances: (params?: { chains?: WalletChain[]; count?: number; startIndex?: number }) => {
-    const query = new URLSearchParams();
-    if (params?.chains?.length) query.set("chains", params.chains.join(","));
-    if (typeof params?.count === "number") query.set("count", String(params.count));
-    if (typeof params?.startIndex === "number") query.set("startIndex", String(params.startIndex));
-    return fetchApi<WalletBalance[]>(`/wallet/balances${query.size ? `?${query.toString()}` : ""}`);
-  },
-  tokenBalances: (params: { chain: WalletTokenChain; index?: number; includeZero?: boolean }) => {
-    const query = new URLSearchParams();
-    query.set("chain", params.chain);
-    if (typeof params.index === "number") query.set("index", String(params.index));
-    if (params.includeZero) query.set("includeZero", "true");
-    return fetchApi<WalletTokenBalance[]>(`/wallet/tokens?${query.toString()}`);
-  },
-  tokenTransactions: (params: {
-    chain: WalletTokenChain;
-    index?: number;
-    limit?: number;
-    tokenAddress?: string;
-    rpcUrl?: string;
-  }) => {
-    const query = new URLSearchParams();
-    query.set("chain", params.chain);
-    if (typeof params.index === "number") query.set("index", String(params.index));
-    if (typeof params.limit === "number") query.set("limit", String(params.limit));
-    if (params.tokenAddress) query.set("tokenAddress", params.tokenAddress);
-    if (params.rpcUrl) query.set("rpcUrl", params.rpcUrl);
-    return fetchApi<WalletTokenTransaction[]>(`/wallet/token-transactions?${query.toString()}`);
-  },
-  transactions: (params: {
-    chain: WalletChain;
-    index?: number;
-    limit?: number;
-    rpcUrl?: string;
-  }) => {
-    const query = new URLSearchParams();
-    query.set("chain", params.chain);
-    if (typeof params.index === "number") query.set("index", String(params.index));
-    if (typeof params.limit === "number") query.set("limit", String(params.limit));
-    if (params.rpcUrl) query.set("rpcUrl", params.rpcUrl);
-    return fetchApi<WalletTransaction[]>(`/wallet/transactions?${query.toString()}`);
-  },
-  send: (payload: {
-    chain: WalletChain;
-    to: string;
-    amount: string;
-    index?: number;
-    memo?: string;
-    rpcUrl?: string;
-    feeRate?: number;
-  }) =>
-    fetchApi<{ chain: WalletChain; txid: string; explorerUrl: string }>("/wallet/send", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  sendToken: (payload: {
-    chain: WalletTokenChain;
-    tokenAddress: string;
-    to: string;
-    amount: string;
-    index?: number;
-    decimals?: number;
-    memo?: string;
-    rpcUrl?: string;
-  }) =>
-    fetchApi<{
-      chain: WalletTokenChain;
-      txid: string;
-      explorerUrl: string;
-      tokenAddress: string;
-    }>("/wallet/send-token", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  ethContractCall: (payload: {
-    contractAddress: string;
-    abi?: string;
-    method: string;
-    methodSignature?: string;
-    args?: unknown[];
-    index?: number;
-    value?: string;
-    gasLimit?: number | string;
-    gasPriceGwei?: string;
-    maxFeePerGasGwei?: string;
-    maxPriorityFeePerGasGwei?: string;
-    nonce?: number;
-    readOnly?: boolean;
-    rpcUrl?: string;
-  }) =>
-    fetchApi<Record<string, unknown>>("/wallet/eth-contract", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  solProgramInstruction: (payload: {
-    programId: string;
-    keys?: WalletInstructionAccount[];
-    accounts?: WalletInstructionAccount[];
-    dataBase64?: string;
-    dataHex?: string;
-    dataUtf8?: string;
-    index?: number;
-    rpcUrl?: string;
-    computeUnitLimit?: number;
-    computeUnitPriceMicroLamports?: number;
-    skipPreflight?: boolean;
-  }) =>
-    fetchApi<{ chain: "sol"; txid: string; explorerUrl: string }>("/wallet/sol-instruction", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  swapEthUniswap: (payload: {
-    tokenOut: string;
-    amountEth?: string;
-    percent?: number;
-    minAmountOut?: string;
-    slippageBps?: number;
-    deadlineSeconds?: number;
-    index?: number;
-    recipient?: string;
-    rpcUrl?: string;
-    dryRun?: boolean;
-  }) =>
-    fetchApi<WalletSwapEthUniswapResult>("/wallet/swap-eth-uniswap", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  priceQuote: (payload: {
-    source?: "auto" | "chainlink" | "pyth" | "jupiter";
-    symbol?: string;
-    pair?: string;
-    feedAddress?: string;
-    feedId?: string;
-    pythFeedId?: string;
-    mint?: string;
-    quoteCurrency?: string;
-    rpcUrl?: string;
-  }) =>
-    fetchApi<{
-      source: "chainlink" | "pyth" | "jupiter";
-      base: string;
-      quote: string;
-      price: string;
-      confidence?: string;
-      publishTime?: string;
-      feedAddress?: string;
-      feedId?: string;
-      mint?: string;
-    }>("/wallet/price", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  swap: (payload: {
-    venue: "uniswap_v2" | "uniswap_v3" | "jupiter" | "uniswap" | "jup" | string;
-    tokenOut?: string;
-    tokenAddress?: string;
-    amountEth?: string;
-    percent?: number;
-    minAmountOut?: string;
-    recipient?: string;
-    feeTier?: number;
-    inputMint?: string;
-    outputMint?: string;
-    amount?: string;
-    amountRaw?: string;
-    index?: number;
-    slippageBps?: number;
-    deadlineSeconds?: number;
-    rpcUrl?: string;
-    wrapUnwrapSol?: boolean;
-    computeUnitPriceMicroLamports?: number;
-    skipPreflight?: boolean;
-    dryRun?: boolean;
-    execute?: boolean;
-    broadcast?: boolean;
-  }) =>
-    fetchApi<{
-      venue: "uniswap_v2" | "uniswap_v3" | "jupiter";
-      chain: "eth" | "sol";
-      from: string;
-      inputToken: string;
-      outputToken: string;
-      amountIn: string;
-      amountInRaw: string;
-      quotedAmountOut: string;
-      quotedAmountOutRaw: string;
-      minAmountOut: string;
-      minAmountOutRaw: string;
-      slippageBps: number;
-      dryRun: boolean;
-      route?: string;
-      routePlan?: Array<{
-        label?: string;
-        ammKey?: string;
-        inputMint?: string;
-        outputMint?: string;
-        inAmount?: string;
-        outAmount?: string;
-      }>;
-      txid?: string;
-      explorerUrl?: string;
-    }>("/wallet/swap", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  endpoints: () =>
-    fetchApi<{
-      ethereum: {
-        wrappedNative: string;
-        dex: Record<string, string>;
-        oracles: {
-          chainlinkFeedRegistry: string;
-          usdDenomination: string;
-          chainlinkUsdFeeds: Record<string, string>;
-          chainlinkBaseAssets: Record<string, string>;
-        };
-      };
-      solana: {
-        nativeMint: string;
-        commonMints: Record<string, string>;
-        programs: Record<string, string>;
-      };
-      services: Record<string, string>;
-    }>("/wallet/endpoints"),
-  dapps: () =>
-    fetchApi<{
-      adapters: Array<{
-        adapter: string;
-        chain: string;
-        write: boolean;
-        description: string;
-      }>;
-      notes: string[];
-    }>("/wallet/dapps"),
-  rpcCall: (payload: {
-    chain: "eth" | "sol";
-    method: string;
-    params?: unknown[];
-    rpcUrl?: string;
-    id?: string | number;
-  }) =>
-    fetchApi<{
-      chain: "eth" | "sol";
-      rpcUrl: string;
-      method: string;
-      id?: string | number;
-      result?: unknown;
-      error?: unknown;
-    }>("/wallet/rpc-call", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  dapp: (payload: { adapter: string; payload?: Record<string, unknown> }) =>
-    fetchApi<unknown>("/wallet/dapp", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  x402: (payload: {
-    url: string;
-    method?: string;
-    headers?: Record<string, string>;
-    body?: unknown;
-    network?: string;
-    maxAmountAtomic?: string;
-    index?: number;
-    timeoutMs?: number;
-    dryRun?: boolean;
-    parseJsonResponse?: boolean;
-  }) =>
-    fetchApi<{
-      url: string;
-      method: string;
-      status: number;
-      paid: boolean;
-      attemptedPayment: boolean;
-      paymentHeaderUsed?: string;
-      paymentRequirement?: {
-        x402Version: number;
-        scheme: string;
-        network: string;
-        amount: string;
-        asset: string;
-        payTo: string;
-        maxTimeoutSeconds: number;
-        extra?: Record<string, unknown>;
-      };
-      settlement?: {
-        success?: boolean;
-        errorReason?: string;
-        errorMessage?: string;
-        payer?: string;
-        transaction?: string;
-        network?: string;
-      };
-      responseHeaders: Record<string, string>;
-      body?: unknown;
-    }>("/wallet/x402", {
-      method: "POST",
-      body: JSON.stringify(payload),
-    }),
-  signMessage: (message: string, chain: WalletChain = "eth", index = 0) =>
-    fetchApi<{ address: string; signature: string }>("/wallet/sign", {
-      method: "POST",
-      body: JSON.stringify({ message, chain, index }),
-    }),
-  deleteWallet: (password?: string) =>
-    fetchApi<{ success: boolean }>("/wallet", {
-      method: "DELETE",
-      body: password ? JSON.stringify({ password }) : undefined,
-    }),
-  setAgentAccess: (enabled: boolean) =>
-    fetchApi<{ success: boolean; enabled: boolean }>("/wallet/agent-access", {
-      method: "PUT",
-      body: JSON.stringify({ enabled }),
-    }),
-  getAgentPolicy: () => fetchApi<WalletAgentPolicy>("/wallet/agent-policy"),
-  updateAgentPolicy: (policy: Partial<WalletAgentPolicy>) =>
-    fetchApi<{ success: boolean; policy: WalletAgentPolicy }>("/wallet/agent-policy", {
-      method: "PUT",
-      body: JSON.stringify(policy),
-    }),
-};
-
+export * from "@/lib/api/wallet";
 export const memoryApi = {
   list: (params?: { agentId?: string; userId?: string; search?: string; limit?: number }) => {
     const query = new URLSearchParams();
@@ -2111,289 +1560,7 @@ export const chatApi = {
   deleteSession: (id: string) => fetchApi<void>("/sessions/" + id, { method: "DELETE" }),
 };
 
-export interface AgentTrajectoryStructure {
-  tools: Array<{
-    name: string;
-    status: string;
-    argumentKeys: string[];
-    resultKeys: string[];
-  }>;
-  response: {
-    hasContent: boolean;
-    hasThinking: boolean;
-    contentKind: "empty" | "text" | "structured";
-  };
-}
-
-export interface AgentGolden {
-  id: string;
-  trajectoryId: string;
-  name: string;
-  description: string | null;
-  tags: string[];
-  assertions: {
-    response?: { type: string };
-    tools: Array<{ index: number; name?: string }>;
-  };
-  baseline: {
-    id: string;
-    sessionId: string;
-    turnIndex: number;
-    agentId: string;
-    provider: string | null;
-    model: string | null;
-    request: {
-      userMessage: ChatMessage;
-      userMessageIndex: number;
-      workspaceDir: string | null;
-    };
-    structure: AgentTrajectoryStructure;
-    createdAt: string;
-  };
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface AgentEvalRun {
-  id: string;
-  goldenId: string;
-  replaySessionId: string | null;
-  status: "running" | "passed" | "failed" | "error";
-  score: number | null;
-  comparison: {
-    equivalent: boolean;
-    score: number;
-    differences: Array<{
-      path: string;
-      expected: unknown;
-      actual: unknown;
-      severity: "error" | "warning";
-    }>;
-  } | null;
-  error: string | null;
-  createdAt: string;
-  completedAt: string | null;
-}
-
-export type ResearchExportFormat =
-  | "cybara_trace"
-  | "trl_sft"
-  | "distillation_sft"
-  | "hf_session_trace"
-  | "prompt_completion"
-  | "long_context";
-
-export interface LabSettings {
-  enabled: boolean;
-  goldenTurnsEnabled: boolean;
-  trajectoryCaptureEnabled: boolean;
-  sanitizeExportsByDefault: boolean;
-  defaultExportFormat: ResearchExportFormat;
-}
-
-export interface ResearchTraceSummary {
-  id: string;
-  sessionId: string;
-  turnIndex: number;
-  agentId: string;
-  provider: string | null;
-  model: string | null;
-  promptPreview: string;
-  responsePreview: string;
-  messageCount: number;
-  toolCallCount: number;
-  toolNames: string[];
-  failedToolCallCount: number;
-  hasObservableReasoning: boolean;
-  observableReasoningCharacters: number;
-  qualityScore: number;
-  qualityFlags: string[];
-  split: "train" | "validation" | "test";
-  createdAt: string;
-}
-
-export interface ResearchTraceStats {
-  total: number;
-  toolCalls: number;
-  failedToolCalls: number;
-  reasoningTraces: number;
-  cleanTraces: number;
-  train: number;
-  validation: number;
-  test: number;
-}
-
-export type IntelligenceTaskDifficulty =
-  | "basic"
-  | "intermediate"
-  | "advanced"
-  | "expert"
-  | "stress";
-
-export interface IntelligenceBenchmarkResult {
-  taskId: string;
-  label: string;
-  category: "instruction" | "reasoning" | "coding" | "transformation" | "tool_use";
-  passed: boolean;
-  score: number;
-  rating?: number;
-  response: string;
-  expected: string;
-  difficulty: IntelligenceTaskDifficulty;
-  weight: number;
-  gradingReason: string;
-  durationMs: number;
-  toolCalls: string[];
-  error: string | null;
-}
-
-export interface IntelligenceBenchmarkRun {
-  id: string;
-  suiteId: string;
-  agentId: string;
-  provider: string | null;
-  model: string | null;
-  status: "running" | "completed" | "cancelled" | "error";
-  score: number;
-  currentTask: number;
-  results: IntelligenceBenchmarkResult[];
-  error: string | null;
-  createdAt: string;
-  completedAt: string | null;
-}
-
-export const evalsApi = {
-  list: () => fetchApi<{ goldens: AgentGolden[]; runs: AgentEvalRun[] }>("/evals"),
-  export: (format: "bundle" | "jsonl", sanitize: boolean) =>
-    fetchApi<{
-      filename: string;
-      mimeType: string;
-      content: string;
-      count: number;
-    }>(`/evals/export?format=${format}&sanitize=${sanitize ? "1" : "0"}`),
-  import: (bundle: unknown) =>
-    fetchApi<{
-      success: boolean;
-      imported: AgentGolden[];
-      count: number;
-      error?: string;
-    }>("/evals/import", { method: "POST", body: JSON.stringify({ bundle }) }),
-  replay: (goldenId: string, payload?: { agentId?: string; modelOverride?: string }) =>
-    fetchApi<{ success: boolean; run: AgentEvalRun; error?: string }>(
-      `/evals/goldens/${goldenId}/replay`,
-      { method: "POST", body: JSON.stringify(payload ?? {}) }
-    ),
-  updateAssertions: (goldenId: string, assertions: AgentGolden["assertions"]) =>
-    fetchApi<{ success: boolean; golden?: AgentGolden; error?: string }>(
-      `/evals/goldens/${goldenId}/assertions`,
-      { method: "PUT", body: JSON.stringify({ assertions }) }
-    ),
-  runSuite: (goldenIds?: string[]) =>
-    fetchApi<{ success: boolean; runs: AgentEvalRun[]; error?: string }>("/evals/run", {
-      method: "POST",
-      body: JSON.stringify(goldenIds ? { goldenIds } : {}),
-    }),
-  deleteGolden: (goldenId: string) =>
-    fetchApi<{ success: boolean }>(`/evals/goldens/${goldenId}`, {
-      method: "DELETE",
-    }),
-};
-
-export const researchApi = {
-  traces: (limit = 200, offset = 0) =>
-    fetchApi<{
-      traces: ResearchTraceSummary[];
-      stats: ResearchTraceStats;
-      total: number;
-      limit: number;
-      offset: number;
-    }>(`/evals/research/traces?limit=${limit}&offset=${offset}`),
-  export: (format: ResearchExportFormat, sanitize: boolean, ids: string[]) => {
-    const params = new URLSearchParams({
-      format,
-      sanitize: sanitize ? "1" : "0",
-    });
-    if (ids.length > 0) params.set("ids", ids.join(","));
-    return fetchApi<{
-      format: ResearchExportFormat;
-      filename: string;
-      mimeType: string;
-      content: string;
-      count: number;
-    }>(`/evals/research/export?${params.toString()}`);
-  },
-  datasetCard: (format: ResearchExportFormat, sanitize: boolean, ids: string[]) => {
-    const params = new URLSearchParams({
-      format,
-      sanitize: sanitize ? "1" : "0",
-    });
-    if (ids.length > 0) params.set("ids", ids.join(","));
-    return fetchApi<{
-      filename: "README.md";
-      mimeType: "text/markdown";
-      content: string;
-      count: number;
-    }>(`/evals/research/card?${params.toString()}`);
-  },
-};
-
-export const benchmarksApi = {
-  list: () =>
-    fetchApi<{
-      suite: {
-        id: string;
-        name: string;
-        description: string;
-        taskCount: number;
-        minRating?: number;
-        maxRating?: number;
-        tasks: Array<{
-          id: string;
-          label: string;
-          category: string;
-          prompt: string;
-          rating?: number;
-          difficulty: IntelligenceTaskDifficulty;
-          weight: number;
-          requiredTool?: string;
-        }>;
-      };
-      runs: IntelligenceBenchmarkRun[];
-    }>("/evals/benchmarks"),
-  run: (agentId: string) =>
-    fetchApi<{
-      success: boolean;
-      run?: IntelligenceBenchmarkRun;
-      error?: string;
-    }>("/evals/benchmarks/run", {
-      method: "POST",
-      body: JSON.stringify({ agentId }),
-    }),
-  export: () =>
-    fetchApi<{
-      filename: string;
-      mimeType: string;
-      content: string;
-      count: number;
-    }>("/evals/benchmarks/export"),
-  manifest: () =>
-    fetchApi<{ filename: string; mimeType: string; content: string }>("/evals/benchmarks/manifest"),
-  cancel: (runId: string) =>
-    fetchApi<{
-      success: boolean;
-      run?: IntelligenceBenchmarkRun;
-      error?: string;
-    }>("/evals/benchmarks/cancel", {
-      method: "POST",
-      body: JSON.stringify({ runId }),
-    }),
-  remove: (runId: string) =>
-    fetchApi<{ success: boolean; error?: string }>("/evals/benchmarks", {
-      method: "DELETE",
-      body: JSON.stringify({ runId }),
-    }),
-};
-
+export * from "@/lib/api/lab";
 export const workspaceOpenApi = {
   targets: (path: string, signal?: AbortSignal) =>
     fetchApi<{
@@ -2413,148 +1580,7 @@ export const dashboardApi = {
   getStats: () => fetchApi<DashboardStats>("/dashboard/stats"),
 };
 
-export interface ComputerUseStatus {
-  available: boolean;
-  command: string;
-  driverSource?:
-    | "env"
-    | "config"
-    | "bundled"
-    | "managed-runtime"
-    | "path"
-    | "known-install-dir"
-    | "default";
-  configuredCommand?: string;
-  platform: string;
-  version?: string;
-  accessibility?: boolean;
-  screenRecording?: boolean;
-  ready: boolean;
-  message: string;
-  installHint?: string;
-  searchedPaths?: string[];
-}
-
-export interface ComputerUseTrajectorySummary {
-  id: string;
-  sessionId: string;
-  status: "recording" | "completed" | "interrupted" | "error";
-  recordVideo: boolean;
-  createdAt: string;
-  updatedAt: string;
-  completedAt?: string;
-  error?: string;
-  replayOf?: string;
-  turnCount: number;
-  screenshotCount: number;
-  clickCount: number;
-  durationMs: number;
-  videoAvailable: boolean;
-}
-
-export interface ComputerUseTrajectoryTurn {
-  index: number;
-  tool: string;
-  arguments: Record<string, unknown>;
-  result: unknown;
-  pid?: number;
-  clickPoint?: { x: number; y: number };
-  timestamp?: string;
-  hasScreenshot: boolean;
-  hasClickImage: boolean;
-  hasAppState: boolean;
-}
-
-export interface ComputerUseTrajectoryDetail extends ComputerUseTrajectorySummary {
-  turns: ComputerUseTrajectoryTurn[];
-}
-
-export interface ComputerUseTrajectorySettings {
-  driverCommand: string;
-  trajectoryCaptureEnabled: boolean;
-  trajectoryVideoEnabled: boolean;
-}
-
-export const computerUseApi = {
-  getStatus: () => fetchApi<ComputerUseStatus>("/computer-use/status"),
-  grantPermissions: () =>
-    fetchApi<{ ok: boolean; message: string }>("/computer-use/permissions/grant", {
-      method: "POST",
-    }),
-  trajectories: () =>
-    fetchApi<{
-      trajectories: ComputerUseTrajectorySummary[];
-      activeId: string | null;
-      settings: ComputerUseTrajectorySettings;
-    }>("/computer-use/trajectories"),
-  trajectory: (id: string) =>
-    fetchApi<{
-      success: boolean;
-      trajectory?: ComputerUseTrajectoryDetail;
-      error?: string;
-    }>(`/computer-use/trajectories/${encodeURIComponent(id)}`),
-  configureTrajectories: (settings: {
-    trajectoryCaptureEnabled?: boolean;
-    trajectoryVideoEnabled?: boolean;
-  }) =>
-    fetchApi<{ success: boolean; settings: ComputerUseTrajectorySettings }>(
-      "/computer-use/trajectories/config",
-      { method: "POST", body: JSON.stringify(settings) }
-    ),
-  exportTrajectories: (ids: string[], includeMedia: boolean, redact: boolean) => {
-    const params = new URLSearchParams({
-      includeMedia: includeMedia ? "1" : "0",
-      redact: redact ? "1" : "0",
-    });
-    if (ids.length > 0) params.set("ids", ids.join(","));
-    return fetchApi<{
-      filename: string;
-      mimeType: string;
-      content: string;
-      count: number;
-    }>(`/computer-use/trajectories/export?${params.toString()}`);
-  },
-  replayTrajectory: (id: string, options?: { delayMs?: number; stopOnError?: boolean }) =>
-    fetchApi<{
-      success: boolean;
-      source: ComputerUseTrajectoryDetail;
-      replay: ComputerUseTrajectoryDetail | null;
-      result: string;
-    }>(`/computer-use/trajectories/${encodeURIComponent(id)}/replay`, {
-      method: "POST",
-      body: JSON.stringify(options ?? {}),
-    }),
-  deleteTrajectory: (id: string) =>
-    fetchApi<{ success: boolean }>(`/computer-use/trajectories/${encodeURIComponent(id)}`, {
-      method: "DELETE",
-    }),
-};
-
-export interface SandboxBrowserStatus {
-  dockerAvailable: boolean;
-  imageBuilt: boolean;
-  running: boolean;
-  cdpPort: number;
-  novncPort: number;
-  cdpUrl: string;
-  novncUrl: string;
-  reason?: string;
-}
-
-export const sandboxBrowserApi = {
-  getStatus: () => fetchApi<SandboxBrowserStatus>("/browser/sandbox/status"),
-  start: () =>
-    fetchApi<{
-      success: boolean;
-      status?: SandboxBrowserStatus;
-      error?: string;
-    }>("/browser/sandbox/start", { method: "POST" }),
-  stop: () =>
-    fetchApi<{ success: boolean; status?: SandboxBrowserStatus }>("/browser/sandbox/stop", {
-      method: "POST",
-    }),
-};
-
+export * from "@/lib/api/computer-use";
 export interface LogPageEntry {
   id: string;
   level: string;
